py-django/pyproject.toml
[tool.poetry]
name = "py-django"
version = "0.1.0"
description = ""
authors = ["Armando Solheiro"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
kombu = "^5.4.2"
gunicorn = "^23.0.0"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
